// frontend/src/types/search.types.ts
export type SearchMode = "CEP" | "ADDRESS";

export interface Address {
  cep: string;
  logradouro: string;
  complemento: string;
  unidade: string;
  bairro: string;
  localidade: string;
  uf: string;
  estado: string;
  regiao: string;
  ibge: string;
  gia: string;
  ddd: string;
  siafi: string;
}

export interface SearchByAddressParams {
  uf: string;
  cidade: string;
  rua: string;
}